% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout_survival.R
\name{checkout_survival}
\alias{checkout_survival}
\title{check the survival analysis}
\usage{
checkout_survival(exp, survivalrank, pcol = 4, symbolcol = 1, path)
}
\arguments{
\item{exp}{This is a transcriptome count matrix,Note that the matrix
requires symbol as column name and sample as row name}

\item{survivalrank}{This is the result calculated by the frist function
in the survival666 package.
However,that does not prevent you from getting results that you could get
using other methods,as long as three conditions are met.
Frist,you need to calculate the p value based on the expression of a
particular phenotype and the level of gene expresion
Second,your results must be large enough,preferably for the entire
transcriptome,otherwise filtering accuracy will be reduced
Finally,you need to eliminate genes that are not significant enough,and if
you are using a regular PC,it is recommended to have no more than 2000 lines}

\item{pcol}{Digtal vector,represents the column of P value in survivalrank}

\item{symbolcol}{Digtal vector,represents the column of symbol in survivalrank}

\item{path}{Path to output file,do not omit '/',example:'d:/file/'.}
}
\description{
The checkout_surviva() is the second function of package survival666.
We use the function to filter the results of super_survival().
}
\examples{
\dontrun{checkout_survival(exp=survivaldata,
survivalrank=survivalrank[1:300,],
pcol=4,
symbolcol=1,
path='/yourdir/')}
}
\seealso{
checkout_survival  \code{\link{super_survival}}
}
