% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic.R
\docType{data}
\name{clinic}
\alias{clinic}
\title{Clinical records of overall survival}
\format{
An object of class \code{data.frame} with 454 rows and 3 columns.
}
\source{
\url{https://xenabrowser.net/datapages/}
}
\usage{
clinic
}
\description{
This is a document that records the overall survival rate.
}
\keyword{datasets}
