% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_pie.R
\name{survival_pie}
\alias{survival_pie}
\title{Visualize the factors that influence survival analysis}
\usage{
survival_pie(exp, survivalrank, target, pcol = 4, symbolcol = 1, path)
}
\arguments{
\item{exp}{This is a transcriptome count matrix,Note that the matrix
requires symbol as column name and sample as row name}

\item{survivalrank}{This is the result calculated by the frist function
in the survival666 package.
However,that does not prevent you from getting results that you could get
using other methods,as long as three conditions are met.
Frist,you need to calculate the p value based on the expression of a
particular phenotype and the level of gene expresion
Second,your results must be large enough,preferably for the entire
transcriptome,otherwise filtering accuracy will be reduced
Finally,you need to eliminate genes that are not significant enough,and if
you are using a regular PC,it is recommended to have no more than 2000 lines}

\item{target}{target gene symbols}

\item{pcol}{Digtal vector,represents the column of P value in survivalrank}

\item{symbolcol}{Digtal vector,represents the column of symbol in survivalrank}

\item{path}{Path to output file,do not omit '/',example:'d:/file/'.}
}
\description{
The checkout_survival() is the third function of survival666.
We use the function to visualize the co-expression interference of a specific
gene in survival analysis.
}
\examples{
\dontrun{survival_pie(exp=survivaldata,
survivalrank=survivalrank[1:300,],
target="EAF2",
pcol=4,
symbolcol=1,
path='/yourdir/')}
}
\seealso{
survival_pie  \code{\link{super_survival}}
}
