% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivaldata.R
\docType{data}
\name{survivaldata}
\alias{survivaldata}
\title{A filtered transcriptome gene expression matrix.}
\format{
An object of class \code{data.frame} with 454 rows and 1712 columns.
}
\source{
\url{https://xenabrowser.net/datapages/}
}
\usage{
survivaldata
}
\description{
The expression matrix was manipulated to contain less than two thousand genes.
Use the symbol as the column name and the sample number as the row name.
The reason of this is that to save space.It is best to use the complete transcriptome
as the expression matrix when using the fundtion super_survival.
}
\keyword{datasets}
