% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalrank.R
\docType{data}
\name{survivalrank}
\alias{survivalrank}
\title{Rank the significance of survival analysis}
\format{
An object of class \code{data.frame} with 1712 rows and 9 columns.
}
\usage{
survivalrank
}
\description{
This data file is actually the output of super_survival function.
The significance P value of survival analysis for each gene in survivaldata
matrix were recorded
}
\keyword{datasets}
